Move progress rows logic in userActivity
