+++ src/services/content.js
@@ -3,7 +3,6 @@
  */
 
 import {
-  fetchAll,
   fetchByRailContentIds,
   fetchMetadata,
   fetchRecent,
@@ -12,14 +11,12 @@ import {
   fetchUpcomingEvents,
   fetchScheduledReleases,
   fetchReturning,
-  fetchLeaving, fetchScheduledAndNewReleases, fetchHierarchy, fetchMethodChildrenIds, jumpToContinueContent
+  fetchLeaving, fetchScheduledAndNewReleases
 } from './sanity.js'
 import {TabResponseType, Tabs, capitalizeFirstLetter} from '../contentMetaData.js'
-import {getAllStartedOrCompleted, getProgressDateByIds, getProgressStateByIds} from "./contentProgress";
 import {fetchHandler} from "./railcontent";
 import {recommendations} from "./recommendations";
-import {fetchPlaylistItems, fetchUserPlaylists} from "./content-org/playlists";
-import {collectionLessonTypes, lessonTypesMapping, progressTypesMapping} from "../contentTypeConfig";
+
 
 export async function getLessonContentRows (brand='drumeo', pageName = 'lessons') {
   let recentContentIds = await fetchRecent(brand, pageName, { progress: 'recent' });
@@ -373,316 +370,6 @@ export async function getRecommendedForYou(brand, rowId = null, {
   return { id: 'recommended', title: 'Recommended For You', items: contents }
 }
 
-export async function getProgressRows({ brand = null, limit = 8 } = {}) {
-  const excludedTypes = new Set([
-    'course-part',
-    'pack-bundle-lesson',
-    'song-tutorial-children',
-    'challenge-part',
-    'pack-bundle',
-    'semester-pack-lesson',
-    'learning-path-lesson',
-    'learning-path-course',
-    'learning-path-level'
-  ]);
-
-  const recentPlaylists = await fetchUserPlaylists(brand, {
-    sort: '-last_progress',
-    limit: limit,
-  });
-  const playlists = recentPlaylists?.data || [];
-  const eligiblePlaylistItems = await getEligiblePlaylistItems(playlists);
-  const playlistEngagedOnContents = eligiblePlaylistItems.map(item => item.last_engaged_on);
-  const playlistsContents = await fetchByRailContentIds(playlistEngagedOnContents, 'progress-tracker');
-  const excludedParents = new Set();
-  for (const item of playlistsContents) {
-    const contentId = item.id ?? item.railcontent_id;
-    excludedParents.add(contentId)
-    const parentIds = item.parent_content_data || [];
-      parentIds.forEach(id => excludedParents.add(id));
-  }
-
-
-  const progressContents = await getAllStartedOrCompleted({onlyIds: false, brand: brand, excludedIds: Array.from(excludedParents) });
-  const contents = await fetchByRailContentIds(Object.keys(progressContents), 'progress-tracker', brand);
-   console.log('rox:: progress from video      ---------------------------', Object.keys(progressContents), contents);
-  const contentsMap = {};
-  contents.forEach(content => {
-    contentsMap[content.railcontent_id] = content;
-  });
-
-  const childToParentMap = {};
-  const parentToChildrenMap = {};
-  const addedParentIds = new Set();
-
-  Object.values(contentsMap).forEach(content => {
-    if (Array.isArray(content.parent_content_data)) {
-      content.parent_content_data.forEach(parentId => {
-        childToParentMap[content.id] = parentId;
-        if (!parentToChildrenMap[parentId]) {
-          parentToChildrenMap[parentId] = [];
-        }
-        parentToChildrenMap[parentId].push(content.id);
-      });
-    }
-  });
-
-  const progressList = [];
-
-  for (const [idStr, progress] of Object.entries(progressContents)) {
-    const id = parseInt(idStr);
-    const content = contentsMap[id];
-
-    if (!content || excludedTypes.has(content.type)) continue;
-
-    const isLeaf = !Array.isArray(content.lessons) || content.lessons === null;
-    const parentId = childToParentMap[id];
-
-    // If it's a child and we haven't yet added its parent
-    if (parentId ) {
-      console.log('rox:: parentId type:', typeof parentId, content);
-      console.log('rox:: All keys in contentsMap:', Object.keys(contentsMap));
-      const parentContent = contentsMap[parentId];
-
-      if (parentContent) {
-        if(!addedParentIds.has(parentId)){
-        addedParentIds.add(parentId);
-}
-        const existing = progressList.find(item => item.id === parentId);
-
-        if (existing) {
-          // Add childIndex (you can also use an array if there are multiple)
-          existing.childIndex = nextId;
-        } else {
-          // Otherwise, push a new object with everything
-          progressList.push({
-            id: parentId,
-            raw: parentContent,
-            state: progress.status,
-            percent: progress.progress,
-            progressTimestamp: progress.last_update * 1000,
-            childIndex: nextId
-          });
-        }
-
-        console.log('rox:: it\'s a child and we haven\'t yet added its parent    ', content ,'gasit')
-
-      }
-      continue; // skip adding the child
-    }
-
-    // If it's a parent (with lessons) and not added yet
-    if (!parentId && !addedParentIds.has(id)) {
-      progressList.push({
-        id,
-        raw: content,
-        state: progress.status,
-        percent: progress.progress,
-        progressTimestamp: progress.last_update * 1000
-      });
-      addedParentIds.add(id);
-    }
-  }
-
- const combined = mergeAndSortItems([...progressList, ...eligiblePlaylistItems], limit);
-  const results = await Promise.all(
-    combined.slice(0, limit).map(item =>
-      item.type === 'playlist'
-        ? processPlaylistItem1(item)
-        : processContentItem1(item)
-    )
-  );
-
-  return {
-    type: TabResponseType.PROGRESS_ROWS,
-    data: results
-  };
-}
-
-function processContentItem1(item) {
-  const data = item.raw;
-  const progress = item.progress;
-  const contentType = getFormattedType(data.type);
-  const status = item.state;
-  let ctaText = 'Continue';
-  if (contentType === 'transcription') ctaText = 'Replay Song';
-  if (contentType === 'lesson') ctaText = status === 'completed' ? 'Revisit Lesson' : 'Continue';
-  if ((contentType === 'pack' || contentType === 'song tutorial' || collectionLessonTypes.includes(contentType)) && status === 'completed') {
-    ctaText = 'View Lessons';
-  }
- // const { status, progress: progressPercent } = progress;
-  const lessonIds = extractLessonIds(item);
-
-  if(data.lesson_count > 0){
-    console.log('rox:: item parinte 1', item)
-
-
-    if(item.childIndex){
-      const nextId = item.childIndex
-      const lessonIndex = data.lessons.findIndex(lesson => lesson.id === nextId);
-      if(progress.status == "completed" && content.type === 'challenge-part')
-      {
-        const lessonIndex = data.lessons.findIndex(lesson => lesson.id === item.childIndex);
-      } elseif(progress.status == "completed" )
-      {
-
-      }
-
-
-      console.log('rox:: item parinte', item, data.lessons, lessonIndex, item.childIndex)
-      if (lessonIndex !== -1) {
-        const lesson = data.lessons[lessonIndex];
-        data.first_incomplete_child = lesson;
-        console.log('rox:: item parinte 21 ',  lesson)
-      }
-    }
-  }
-  return {
-    id: item.id,
-    progressType: 'content',
-    header: contentType,
-    body: {
-      progressPercent: item.percent,
-      thumbnail: item.raw.thumbnail,
-      title: item.raw.title,
-      subtitle: !item.raw.child_count || item.raw.lesson_count === 1
-                   ? `${item.raw.difficulty_string} • ${item.raw.artist_name}`
-                   : `${item.raw.completed_children} of ${item.raw.lesson_count ?? item.raw.child_count} Lessons Complete`
-    },
-    cta: {
-      text: ctaText,
-      action: {
-        type: data.type,
-        brand: data.brand,
-        id: data.id,
-        slug: data.slug,
-        child: data.first_incomplete_child
-                ? {
-            id: data.first_incomplete_child.id,
-            type: data.first_incomplete_child.type,
-            brand: data.first_incomplete_child.brand,
-            slug: data.first_incomplete_child.slug,
-            child: data.second_incomplete_child
-                  ? {
-                id: data.second_incomplete_child.id,
-                type: data.second_incomplete_child.type,
-                brand: data.second_incomplete_child.brand,
-                slug: data.second_incomplete_child.slug
-              }
-                  : null
-          }
-                : null
-      }
-    },
-    progressTimestamp: item.progressTimestamp
-  };
-  return {
-    ...item,
-    type: getFormattedType(item.type),
-    progressTimestamp: item.progressTimestamp,
-  };
-}
-
-async function processPlaylistItem1(item) {
-  const playlist = item.raw;
-  const progressOnItems = await getProgressStateByIds(playlist.items.map(a => a.content_id));
-  const allItemsCompleted = item.raw.items.every(i => {
-    const itemId = i.content_id;
-    const progress = progressOnItems[itemId];
-    return progress && progress === 'completed';
-  });
-
-  let nextItem = playlist.items[0] ?? null;
-  if (!allItemsCompleted) {
-    const lastItemProgress = progressOnItems[playlist.last_engaged_on];
-    const index = playlist.items.findIndex(i => i.content_id  === playlist.last_engaged_on);
-    if (lastItemProgress === 'completed') {
-      nextItem = playlist.items[index + 1] ?? nextId;
-    } else {
-      nextItem = playlist.items[index] ?? nextId;
-    }
-  }
-
-  return {
-    id:                playlist.id,
-    progressType:      'playlist',
-    header:            'playlist',
-    body:              {
-      first_items_thumbnail_url: playlist.first_items_thumbnail_url,
-      title:                     playlist.name,
-      subtitle:                  `${playlist.duration_formated} • ${playlist.total_items} items • ${playlist.likes} likes • ${playlist.user.display_name}`
-    },
-    progressTimestamp: item.progressTimestamp,
-    cta:               {
-      text:   'Continue',
-      action: {
-        brand:  playlist.brand,
-        id:     playlist.id,
-        itemId: nextItem.id,
-        type:   'playlists',
-      }
-    }
-  }
-}
-const getFormattedType = type => {
-  for (const [key, values] of Object.entries(progressTypesMapping)) {
-    if (values.includes(type)) {
-      return key;
-    }
-  }
-  return null;
-};
-
-function traverse(lessons) {
-  for (const item of lessons) {
-    if (item.id) {
-      ids.push(item.id);
-    }
-    if (item.lessons) {
-      traverse(item.lessons); // Recursively handle nested lessons
-    }
-  }
-}
-
-function extractLessonIds(data) {
-  const ids = [];
-  function traverse(lessons) {
-    for (const item of lessons) {
-      if (item.id) {
-        ids.push(item.id);
-      }
-      if (item.lessons) {
-        traverse(item.lessons); // Recursively handle nested lessons
-      }
-    }
-  }
-  if (data.raw && Array.isArray(data.raw.lessons)) {
-    traverse(data.raw.lessons);
-  }
-
-  return ids;
-}
-
-
-async function getEligiblePlaylistItems(playlists) {
-  return Promise.all(
-    playlists
-      .filter(p => p.last_progress && p.last_engaged_on)
-      .map(async p => ({
-        type: 'playlist',
-        progressTimestamp: new Date(p.last_progress).getTime(),
-        last_engaged_on:p.last_engaged_on,
-        raw: p
-      }))
-  );
-}
-
-function mergeAndSortItems(items, limit) {
-  return items
-    .filter(item => typeof item.progressTimestamp === 'number' && item.progressTimestamp > 0)
-    .sort((a, b) => b.progressTimestamp - a.progressTimestamp)
-    .slice(0, limit + 5);
-}
